Remove no-longer needed BOOST_INCLUDE_DIR
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@26313 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/operators/FiveFoldTetSubdivision/Branch.h
@@ -1,7 +1,6 @@
 #ifndef _BRANCH_H_
 #define _BRANCH_H_
 
-#include <cstdint>
 #include <algorithm>
 #include <cassert>
 #include <iostream>
@@ -10,7 +9,7 @@
 class Branch
 {
     public:
-        typedef uint64_t IdxT;
+        typedef unsigned long long IdxT;
         typedef double ValueT;
         IdxT extremum;
         ValueT extremumVal;
